requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.28
pydantic==2.6.4
python-dotenv==1.0.1
pymysql==1.1.0
cryptography==42.0.5  # Versão recente no momento
werkzeug==3.0.1
python-dateutil==2.9.0.post0